Add: Seeders file to all tables

+++ database/migrations/2024_03_20_063856_create_preferences_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
@@ -13,16 +12,16 @@ return new class extends Migration
     {
         Schema::create('preferences', function (Blueprint $table) {
             $table->id();
-            $table->double('preferenceMood', 8, 2)->default(0.0); // 0 -> sad, 1 -> happy
-            $table->double('preferenceActivity', 8, 2)->default(0.0); // 0 -> chilling, 1 -> working
-            $table->double('preferenceCoffeeTemperature', 8, 2)->default(0.0); // 0 -> cold, 1 -> hot
-            $table->double('preferenceCoffeeSweetness', 8, 2)->default(0.0); // 0 -> not sweet, 1 -> sweet
-            $table->double('preferenceCoffeeMilkness', 8, 2)->default(0.0); // 0 -> no milk, 1 -> with milk
-            $table->double('preferenceCoffeePrice', 8, 2)->default(0.0); // 0 -> cheap, 1 -> pricy
-            $table->double('preferenceCoffeeMilkType', 8, 2)->default(0.0); // 0 -> almond milk, 1 -> cow milk
-            $table->double('preferenceCoffeeDrinkType', 8, 2)->default(0.0); // 0 -> pure coffee, 1 -> variant coffee
-            $table->double('preferenceCoffeeAcidityLevel', 8, 2)->default(0.0); // 0 -> not acid,  1 -> acid
-            $table->double('preferenceCoffeeStrengthLevel', 8, 2)->default(0.0); // 0 -> not strong, 1 -> strong
+            $table->integer('preferenceMood')->default(0); // 0 -> sad, 1 -> happy
+            $table->integer('preferenceActivity')->default(0); // 0 -> chilling, 1 -> working
+            $table->integer('preferenceCoffeeTemperature')->default(0); // 0 -> cold, 1 -> hot
+            $table->integer('preferenceCoffeeSweetness')->default(0); // 0 -> not sweet, 1 -> sweet, stepper
+            $table->integer('preferenceCoffeeMilkness')->default(0); // 0 -> no milk, 1 -> with milk
+            $table->integer('preferenceCoffeePrice')->default(0); // 0 -> cheap, 1 -> pricy
+            $table->integer('preferenceCoffeeMilkType')->default(0); // 0 -> almond milk, 1 -> cow milk
+            $table->integer('preferenceCoffeeDrinkType')->default(0); // 0 -> pure coffee, 1 -> variant coffee
+            $table->integer('preferenceCoffeeAcidityLevel')->default(0); // 0 -> not acid,  1 -> acid
+            $table->integer('preferenceCoffeeStrengthLevel')->default(0); // 0 -> not strong, 1 -> strong
             $table->timestamps();
         });
     }
